Feat(codereview): CodeReview Comments API - add CodeReview Comments API - add CodeReview Comments Test code - refactory codes

+++ owl-back/src/main/java/com/project/owlback/codereview/repository/CodeCommentLikeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 import java.sql.SQLException;
 
-public interface CodeCommentLikeRepository extends JpaRepository<CodeCommentLike, Integer> {
+public interface CodeCommentLikeRepository extends JpaRepository<CodeCommentLike, Long> {
     int countByUserIdAndCodeCommentId(int userId, int codeCommentId) throws SQLException;
 }
